refactor: Moved Notion Renderer to Common Component
