+++ src/app/blog/[type]/[id]/_components/index.ts
@@ -1,4 +1,4 @@
 export * from './BlogNavigationButtons'
-export * from './NotionRenderer'
+export * from '~/utils/notion/NotionRenderer'
 export * from './DevPostView'
 export * from './ProjectPostView'
